refactor(lotto): 도메인 및 입력 관련 패키지 구조 개편
lotto의 파일들을 lotto.domain 및 lotto.input 패키지로 분리하여 책임 명확화.
Lotto, Money 등 주요 클래스를 domain 패키지로 이동하고, 입력 뷰 관련 코드를 input 패키지로 이동.
관련 테스트 코드의 패키
