+++ src/main/kotlin/lotto/domain/BonusBallLotto.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 class BonusBallLotto(
     winLotto: WinLotto,
