+++ src/main/kotlin/lotto/domain/Lotto.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 @JvmInline
 value class Lotto(val numbers: List<LottoNumber>) {
